v1.2 done bug fixed checks updated in the next version is expected: flag supporting visualisation and much more

+++ includes/libft_structs.h
@@ -40,7 +40,7 @@ typedef struct				s_list
 
 typedef struct				s_heap
 {
-	int						key;
+	size_t					key;
 	int						dist;
 	struct s_heap			*left;
 	struct s_heap			*right;
